Tracker updates
Updated tracker so it uses the website to find the most recent achievement. Tons of logic and shit to figure it out. Still uses the api for some tasks but it updates much faster. Front end works the same exact way. Still need to fix the case when the achievement is earned during the same minute.